[feat] 조회 재행동 시 횟수에 따라 가중치 차등 업데이트 처리

+++ src/main/java/com/dto/project/domain/product/dto/ProductResponse.java
@@ -22,7 +22,7 @@ public class ProductResponse {
     private String origin;
     private String specInfo;
     private String altText;
-    private Integer weightScore;
+    private Double weightScore;
     private String status;
 
     public ProductResponse(Product product) {
