+++ src/main/java/com/dto/project/domain/weighting/dto/MetaTagWeightLogResponse.java
@@ -14,7 +14,7 @@ public class MetaTagWeightLogResponse {
     private Long memberId;
     private Long metaTagId;
     private WeightLogType actionType;
-    private Integer appliedWeight;
+    private Double appliedWeight;
     private Long referenceId;
     private LocalDateTime eventTimeStamp;
     private LocalDateTime createdAt;
